Update permit add other utility and update claim calculate

+++ forms/apv.aspx.cs
@@ -1489,8 +1489,8 @@ namespace onlineLegalWF.forms
 
                 //check corditon deviation claim
                 float deviation = 0;
-                float cal_iar_uatc = (!string.IsNullOrEmpty(xiar_uatc) ? float.Parse(int.Parse(xiar_uatc, NumberStyles.AllowThousands).ToString()) : 0);
-                float cal_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(int.Parse(xiar_pfc, NumberStyles.AllowThousands).ToString()) : 0);
+                float cal_iar_uatc = (!string.IsNullOrEmpty(xiar_uatc) ? float.Parse(float.Parse(xiar_uatc, NumberStyles.AllowThousands).ToString()) : 0);
+                float cal_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(float.Parse(xiar_pfc, NumberStyles.AllowThousands).ToString()) : 0);
                 float int_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(xiar_pfc, NumberStyles.AllowThousands) : 0);
                 deviation = cal_iar_uatc / cal_iar_pfc;
                 if (int_iar_pfc <= 100000)
