+++ frmInsurance/InsuranceClaim.aspx.cs
@@ -114,8 +114,8 @@ namespace onlineLegalWF.frmInsurance
 
                 //check corditon deviation claim
                 float deviation = 0;
-                float cal_iar_uatc = (!string.IsNullOrEmpty(xiar_uatc) ? float.Parse(int.Parse(xiar_uatc, NumberStyles.AllowThousands).ToString()) : 0);
-                float cal_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(int.Parse(xiar_pfc, NumberStyles.AllowThousands).ToString()) : 0);
+                float cal_iar_uatc = (!string.IsNullOrEmpty(xiar_uatc) ? float.Parse(float.Parse(xiar_uatc, NumberStyles.AllowThousands).ToString()) : 0);
+                float cal_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(float.Parse(xiar_pfc, NumberStyles.AllowThousands).ToString()) : 0);
                 float int_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(xiar_pfc, NumberStyles.AllowThousands) : 0);
                 deviation = cal_iar_uatc / cal_iar_pfc;
                 if (int_iar_pfc <= 100000)
@@ -415,8 +415,8 @@ namespace onlineLegalWF.frmInsurance
 
             //check corditon deviation claim
             float deviation = 0;
-            float cal_iar_uatc = (!string.IsNullOrEmpty(xiar_uatc) ? float.Parse(int.Parse(xiar_uatc, NumberStyles.AllowThousands).ToString()) : 0);
-            float cal_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(int.Parse(xiar_pfc, NumberStyles.AllowThousands).ToString()) : 0);
+            float cal_iar_uatc = (!string.IsNullOrEmpty(xiar_uatc) ? float.Parse(float.Parse(xiar_uatc, NumberStyles.AllowThousands).ToString()) : 0);
+            float cal_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(float.Parse(xiar_pfc, NumberStyles.AllowThousands).ToString()) : 0);
             float int_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(xiar_pfc, NumberStyles.AllowThousands) : 0);
             deviation = cal_iar_uatc / cal_iar_pfc;
             if (int_iar_pfc <= 100000)
@@ -787,8 +787,8 @@ namespace onlineLegalWF.frmInsurance
 
             //check corditon deviation claim
             float deviation = 0;
-            float cal_iar_uatc = (!string.IsNullOrEmpty(xiar_uatc) ? float.Parse(int.Parse(xiar_uatc, NumberStyles.AllowThousands).ToString()) : 0);
-            float cal_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(int.Parse(xiar_pfc, NumberStyles.AllowThousands).ToString()) : 0);
+            float cal_iar_uatc = (!string.IsNullOrEmpty(xiar_uatc) ? float.Parse(float.Parse(xiar_uatc, NumberStyles.AllowThousands).ToString()) : 0);
+            float cal_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(float.Parse(xiar_pfc, NumberStyles.AllowThousands).ToString()) : 0);
             float int_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(xiar_pfc, NumberStyles.AllowThousands) : 0);
             deviation = cal_iar_uatc / cal_iar_pfc;
             if (int_iar_pfc <= 100000)
@@ -1062,8 +1062,8 @@ namespace onlineLegalWF.frmInsurance
 
                 //check corditon deviation claim
                 float deviation = 0;
-                float cal_iar_uatc = (!string.IsNullOrEmpty(xiar_uatc) ? float.Parse(int.Parse(xiar_uatc, NumberStyles.AllowThousands).ToString()) : 0);
-                float cal_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(int.Parse(xiar_pfc, NumberStyles.AllowThousands).ToString()) : 0);
+                float cal_iar_uatc = (!string.IsNullOrEmpty(xiar_uatc) ? float.Parse(float.Parse(xiar_uatc, NumberStyles.AllowThousands).ToString()) : 0);
+                float cal_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(float.Parse(xiar_pfc, NumberStyles.AllowThousands).ToString()) : 0);
                 float int_iar_pfc = (!string.IsNullOrEmpty(xiar_pfc) ? float.Parse(xiar_pfc, NumberStyles.AllowThousands) : 0);
                 deviation = cal_iar_uatc / cal_iar_pfc;
                 if (int_iar_pfc <= 100000)
